pyproject.toml
[project]
name = "school_diary"
version = "0.1.0"
description = "業務Webアプリ共通土台の検証用プロジェクト"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "BizApp Team", email = "admin@example.com" }]
requires-python = "==3.12.*"
dependencies = [
    "argon2-cffi==25.1.0",
    "crispy-bootstrap5==2025.6",
    "django==5.1.12",
    "django-allauth[mfa]==65.11.1",
    "django-anymail[amazon-ses]==13.1",
    "django-crispy-forms==2.4",
    "django-environ==0.12.0",
    "django-model-utils==5.0.0",
    "django-storages[s3]==1.14.6",
    "gunicorn==23.0.0",
    "pillow==11.3.0",
    "psycopg[c]==3.2.9",
    "python-slugify==8.0.4",
    "whitenoise==6.9.0",
    "django-simple-history~=3.5.0",
    "django-fsm~=2.8.1",
    "djangorestframework~=3.15.2",
    "bleach==6.1.0",
    "markdown==3.6",
    "django-jazzmin>=3.0.1",
]

[project.optional-dependencies]
# 開発用の依存関係
dev = [
    "coverage==7.10.6",
    "django-coverage-plugin==3.1.1",
    "django-debug-toolbar==6.0.0",
    "django-extensions==4.1",
    "django-stubs[compatible-mypy]==5.2.2",
    "djlint==1.36.4",
    "factory-boy==3.3.2",
    "ipdb==0.13.13",
    "mypy==1.17.1",
    "pre-commit==4.3.0",
    "psycopg[c]==3.2.9",
    "pytest==8.4.2",
    "pytest-django==4.11.1",
    "pytest-sugar==1.1.1",
    "ruff==0.12.12",
    "sphinx==8.2.3",
    "sphinx-autobuild==2025.8.25",
    "watchfiles==1.1.0",
    "werkzeug[watchdog]==3.1.3",
]

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = ["tests.py", "test_*.py"]

# ==== Coverage ====
[tool.coverage.run]
include = ["school_diary/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[mypy.plugins.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

# ==== Ruff ====
[tool.ruff]
line-length = 120  # Django標準（Instagram, Sentry, GitLab等で採用）
target-version = "py312"
# これらのパターンに一致するディレクトリ/ファイルはチェック対象外にする
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
    "media/*",
    "verify_*.py",    # 検証スクリプト（本番コードではない）
    "create_*.py",    # データ作成スクリプト（本番コードではない）
    "manual_*.py",    # 手動実行スクリプト（本番コードではない）
]

[tool.ruff.lint]
# 有効にするルールのリスト
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    #"BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    'ISC',
    "ICN",
    "G",
    'INP',
    'PIE',
    "T20",
    'PYI',
    'PT',
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TC",
    "INT",
    "PTH",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "PERF",
    "RUF",
]
# 無効にする（無視する）特定のルール
ignore = [
    # 既存ルール
    "S101",   # assertの使用を許可
    "RUF012", # ClassVarの警告を無視
    "SIM102", # ネストしたif文を許可

    # 日本語業務アプリケーション用設定（業界標準：Instagram, Sentry, GitLab等）
    "RUF001",  # 文字列内の全角文字を許可（日本語UI）
    "RUF002",  # docstring内の全角文字を許可（日本語ドキュメント）
    "RUF003",  # コメント内の全角文字を許可（日本語コメント）

    # 開発ツール用設定
    "T201",    # print文を許可（スクリプト・検証ツール用）
    "E501",    # 行の長さはline-length=120で制御

    # 技術的負債（AWS Phase完了後に対応予定）
    "ERA001",   # コメントアウトコード（55件、保守性改善）
    "PLC0415",  # トップレベル外import（99件、保守性改善）
    "PLR2004",  # マジックナンバー（63件、可読性改善）
    "G004",     # logging f-string（24件、スタイル改善）
    "SLF001",   # private member access（23件、django-simple-history標準パターン）
    "DTZ011",   # date.today() timezone（13件、将来対応）
    "DTZ001",   # datetime without timezone（3件、将来対応）
    "E402",     # module import位置（5件、スタイル改善）
    "DJ012",    # django unordered body（6件、スタイル改善）
    "DJ001",    # django nullable string field（5件、設計判断）
    "FBT001",   # boolean positional argument（5件、スタイル改善）
    "FBT002",   # boolean default value（6件、スタイル改善）
    "SIM105",   # suppressible exception（5件、スタイル改善）
    "TRY300",   # try-consider-else（5件、スタイル改善）
    "TRY301",   # raise-within-try（5件、スタイル改善）
    "C901",     # complex structure（3件、リファクタリング対象）
    "PLR0912",  # too many branches（3件、リファクタリング対象）
    "PLR0913",  # too many arguments（2件、リファクタリング対象）
    "PLR0915",  # too many statements（2件、リファクタリング対象）
    "PLR0911",  # too many returns（1件、リファクタリング対象）
    "PERF401",  # manual list comprehension（4件、最適化対象）
    "PERF403",  # manual dict comprehension（1件、最適化対象）
    "PD901",    # pandas df variable name（6件、スタイル改善）
    "S308",     # mark_safe usage（6件、Django標準パターン）
    "S105",     # hardcoded password string（1件、テストコード）
    "S110",     # try-except-pass（1件、意図的）
    "DJ007",    # django ModelForm.Meta.fields（2件、スタイル改善）
    "N806",     # non-lowercase variable（2件、定数パターン）
    "PTH123",   # builtin open（2件、Path対応は将来）
    "RUF005",   # collection literal concatenation（1件、スタイル改善）
    "PGH003",   # blanket type ignore（1件、必要な無視）
    "PLC0206",  # dict index missing items（1件、スタイル改善）
    "TRY401",   # verbose log message（1件、スタイル改善）
]
# 安全でない可能性のある自動修正を有効にするルール
#extend-unsafe-fixes = ["UP038"]

# ファイルごとの無視設定（テストコード用の緩和、業界標準）
[tool.ruff.lint.per-file-ignores]
"**/test_*.py" = ["S106", "PT009"]  # テストコード：ハードコードパスワード、unittest assertion許可
"**/tests.py" = ["S106", "PT009"]
"**/conftest.py" = ["S106"]
"**/commands/*.py" = ["S106"]  # management commands：テスト用パスワード許可

[tool.ruff.lint.isort]
# import文を一行に強制する
force-single-line = true

[tool.setuptools]
packages = ["school_diary", "kits"]

[dependency-groups]
dev = [
    "beautifulsoup4>=4.14.2",
    "pytest-cov>=7.0.0",
]
